Move 'report_processing_qc' function into 'make_fastqs_cmd' and remove the 'qc/processing' module.
